requirements.txt
pdfminer.six==20220319
pymorphy2==0.9.1
scikit-learn==1.0.2
openpyxl==3.0.9
pillow==9.1.0
pytesseract==0.3.9
wand==0.6.7
pdf2image==1.16.0
poppler-utils==0.1.0